+++ data/sqlite/csv.go
@@ -20,13 +20,16 @@ func ExportCSV(ctx context.Context, db *sql.DB, gofiID int, csvSeparator rune, c
 	2. write row by row in a csv (include headers)
 	*/
 	q := ` 
-		SELECT id, year, month, day,
+		SELECT fT.id, year, month, day, fT.mode,
 			account, product, priceIntx100, category, 
-			commentInt, commentString, checked, dateChecked
-		FROM financeTracker
-		WHERE gofiID = ?
+			commentInt, commentString, checked, dateChecked, 
+			IFNULL(lB.name, '')
+		FROM financeTracker AS fT
+			LEFT JOIN specificRecordsByMode AS sR ON fT.gofiID = sR.gofiID AND sR.idFinanceTracker = fT.id
+			LEFT JOIN lenderBorrower AS lB ON lB.gofiID = sR.gofiID AND sR.idLenderBorrower = lB.id
+		WHERE fT.gofiID = ?
 			AND exported = 0
-		ORDER BY id
+		ORDER BY fT.id
 		LIMIT 10000;
 	`
 	rows, err := db.QueryContext(ctx, q, gofiID)
@@ -45,8 +48,8 @@ func ExportCSV(ctx context.Context, db *sql.DB, gofiID int, csvSeparator rune, c
 		nbRows += 1
 		if nbRows == 1 {
 			//write csv headers
-			row = []string{"𫝀é ꮖꭰ", "Date",
-				"Account", "Product", "PriceStr", "Category",
+			row = []string{"𫝀é ꮖꭰ", "Date", "Mode",
+				"Account", "Product", "PriceStr", "Category", "ThirdParty",
 				"CommentInt", "CommentString", "Checked", "DateChecked", "Exported",
 				"."} //keeping a column at the end will handle the LF and CRLF cases
 			if err := w.Write(row); err != nil {
@@ -56,11 +59,12 @@ func ExportCSV(ctx context.Context, db *sql.DB, gofiID int, csvSeparator rune, c
 		}
 		var ft appdata.FinanceTracker
 		var successfull bool
-		var unsuccessfullReason string
+		var unsuccessfullReason, lbName string
 		if err := rows.Scan(
-			&ft.ID, &ft.DateDetails.Year, &ft.DateDetails.Month, &ft.DateDetails.Day,
+			&ft.ID, &ft.DateDetails.Year, &ft.DateDetails.Month, &ft.DateDetails.Day, &ft.Mode,
 			&ft.Account, &ft.Product, &ft.PriceIntx100, &ft.Category,
 			&ft.CommentInt, &ft.CommentString, &ft.Checked, &ft.DateChecked,
+			&lbName,
 		); err != nil {
 			log.Fatal(err)
 		}
@@ -70,8 +74,8 @@ func ExportCSV(ctx context.Context, db *sql.DB, gofiID int, csvSeparator rune, c
 			ft.Date = "ERROR " + unsuccessfullReason
 		}
 
-		row = []string{strconv.Itoa(ft.ID), ft.Date,
-			ft.Account, ft.Product, ft.FormPriceStr2Decimals, ft.Category,
+		row = []string{strconv.Itoa(ft.ID), ft.Date, strconv.Itoa(ft.Mode),
+			ft.Account, ft.Product, ft.FormPriceStr2Decimals, ft.Category, lbName,
 			strconv.Itoa(ft.CommentInt), ft.CommentString, strconv.FormatBool(ft.Checked), ft.DateChecked, "true", "."}
 		if err := w.Write(row); err != nil {
 			fmt.Printf("row error 2: %v\n", row)
@@ -159,26 +163,29 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 		return stringList, true
 	}
 
+	var lb appdata.LendBorrow
 	var ft appdata.FinanceTracker
-	var lineInfo, unsuccessfullReason, controlEncoding, controlLastValidColumn, validControlEncodingUTF8, validControlEncodingUTF8withBOM string
+	var iDlistInt []int
+	var lineInfo, lineInfoTemp, unsuccessfullReason, controlEncoding, controlLastValidColumn, validControlEncodingUTF8, validControlEncodingUTF8withBOM string
 	var successfull bool
-	var flagErr int = 0
+	var flagErr, flagErrTemp int = 0, 0
 	ft.GofiID = gofiID
 	stringList += "𫝀é ꮖꭰ;Date;CommentInt;Checked;exported;NewID;Updated;\n"
 	for index, row := range rows {
 		if index == 0 { //control UTF-8 on headers
 			totalRows := len(row)
-			if totalRows != 12 {
+			if totalRows != 14 {
 				stringList =
 					"IMPORTATION ANNULEE.\n" +
 						"ERREUR sur le nombre de colonnes du fichier.\n\n" +
-						"INFO: total " + strconv.Itoa(totalRows) + " colonnes sur un attendu de 12.\n" +
+						"INFO: total " + strconv.Itoa(totalRows) + " colonnes sur un attendu de 14.\n" +
 						"Un exemple de données d'import valide est disponible plus bas sur cette page."
 				errorBool = true
+				fmt.Printf("INFO: total %v colonnes sur un attendu de 14.\n", strconv.Itoa(totalRows))
 				break //stop
 			}
 			controlEncoding = row[0]
-			controlLastValidColumn = row[10]
+			controlLastValidColumn = row[12]
 			validControlEncodingUTF8 = "𫝀é ꮖꭰ"              //UTF-8
 			validControlEncodingUTF8withBOM = "\ufeff𫝀é ꮖꭰ" //UTF-8 with BOM
 			if (controlEncoding == validControlEncodingUTF8 || controlEncoding == validControlEncodingUTF8withBOM) &&
@@ -190,7 +197,7 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 				stringList =
 					"IMPORTATION ANNULEE.\n" +
 						"ERREUR sur la dernière colonne du fichier.\n\n" +
-						"INFO: 11eme colonne = 'Exported'\n" +
+						"INFO: 13eme colonne = 'Exported'\n" +
 						"Un exemple de données d'import valide est disponible plus bas sur cette page."
 				errorBool = true
 				break //stop
@@ -220,9 +227,11 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 			} else if ft.ID < 0 {
 				// DELETE is actually an UPDATE with empty data
 				lineInfo += "DELETE" + row[0] + ";1999-12-31;;checked true;exported false;"
+				ft.Date = "1999-12-31"
 				ft.DateDetails.Year = 1999
 				ft.DateDetails.Month = 12
 				ft.DateDetails.Day = 31
+				ft.Mode = 0
 				ft.Account = "-"
 				ft.Product = "DELETED LINE"
 				ft.PriceIntx100 = 0
@@ -243,24 +252,41 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 				stringList += lineInfo + "\n"
 				continue //skip this row because wrong date format
 			}
+			ft.Date, successfull, unsuccessfullReason = ConvertDateIntToStr(ft.DateDetails.Year, ft.DateDetails.Month, ft.DateDetails.Day, "EN", "-")
+			if !successfull {
+				ft.Date = "ERROR " + unsuccessfullReason
+			}
 
-			ft.Account = row[2]
-			ft.Product = row[3]
-			ft.FormPriceStr2Decimals = row[4]
+
+			ft.Mode, err = strconv.Atoi(row[2])
+			if err != nil {
+				ft.Mode = 0
+			}
+			ft.Account = row[3]
+			ft.Product = row[4]
+			ft.FormPriceStr2Decimals = row[5]
 			ft.PriceIntx100 = ConvertPriceStrToInt(ft.FormPriceStr2Decimals, csvDecimalDelimiter)
+			if ft.PriceIntx100 < 0 {
+				ft.PriceDirection = "expense"
+				ft.FormPriceStr2Decimals = ft.FormPriceStr2Decimals[1:] //remove the first "-"
+			} else {
+				ft.PriceDirection = "gain"
+				// ft.FormPriceStr2Decimals = ft.FormPriceStr2Decimals
+			}
 
-			ft.Category = row[5]
-			ft.CommentInt, err = strconv.Atoi(row[6])
+			ft.Category = row[6]
+			// ThirdParty = row[7] //handled bellow in lb var
+			ft.CommentInt, err = strconv.Atoi(row[8])
 			if err != nil {
 				ft.CommentInt = 0
 				lineInfo += "comment i 0;"
 			} else {
 				lineInfo += ";"
 			}
-			ft.CommentString = row[7]
+			ft.CommentString = row[9]
 
 			// Checked
-			ft.Checked, err = strconv.ParseBool(row[8])
+			ft.Checked, err = strconv.ParseBool(row[10])
 			if err != nil {
 				ft.Checked = false
 				lineInfo += "checked 0;"
@@ -270,8 +296,8 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 
 			// DateChecked
 			ft.DateChecked = "9999-12-31"
-			if len(row[9]) == 10 {
-				yearInt, monthInt, dayInt, successfull, _ := ConvertDateStrToInt(row[9], dateFormat, dateSeparator)
+			if len(row[11]) == 10 {
+				yearInt, monthInt, dayInt, successfull, _ := ConvertDateStrToInt(row[11], dateFormat, dateSeparator)
 				// fmt.Println("---------------")
 				// fmt.Printf("ft.DateChecked: %v\n", ft.DateChecked)
 				// fmt.Printf("yearInt %v, monthInt %v, dayInt %v, successfull %v, unsuccessfullReason %v\n", yearInt, monthInt, dayInt, successfull, unsuccessfullReason)
@@ -291,38 +317,39 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 		}
 		if ft.ID == 0 {
 			// INSERT
-			exec, err := db.ExecContext(ctx, `
-				INSERT INTO financeTracker (gofiID, dateIn, year, month, day, account, product, priceIntx100, category,
-					commentInt, commentString, checked, dateChecked, exported)
-				VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,0);
-				`,
-				ft.GofiID, ft.Date, ft.DateDetails.Year, ft.DateDetails.Month, ft.DateDetails.Day, ft.Account, ft.Product, ft.PriceIntx100, ft.Category,
-				ft.CommentInt, ft.CommentString, ft.Checked, ft.DateChecked,
-			)
-			if err != nil {
-				lineInfo += "error1;false;"
-				fmt.Printf("error1: %#v\n", err)
-				flagErr += 1
-			} else {
-				rowID, err := exec.LastInsertId()
-				if err != nil {
-					lineInfo += "error2;false;"
-					fmt.Printf("error2: %#v\n", err)
+			if ft.Mode == 0 {
+				idFT, isErr, _, info := handleFTinsertCSV(ctx, &ft, csvSeparator, csvDecimalDelimiter, dateFormat, dateSeparator)
+				if isErr {
+					lineInfo += "error1;false;"
+					fmt.Printf("error1: %v\n", info)
 					flagErr += 1
+					// return appdata.RenderAPIorUI(w, r, isFrontRequest, false, false, httpCode, info, "")
 				} else {
-					lineInfo += strconv.FormatInt(rowID, 10) + ";true;"
+					lineInfo += strconv.FormatInt(idFT, 10) + ";true;"
 				}
+			} else if ft.Mode >= 1 && ft.Mode <= 4 {
+				lb.FT = ft
+				lb.ModeInt = ft.Mode
+				lb.CreateLenderBorrowerName = row[7]
+				lb.Who = row[7]			
+				lineInfoTemp, flagErrTemp = handleLBinsertCSV(ctx, &lb, csvSeparator, csvDecimalDelimiter, dateFormat, dateSeparator)
+				lineInfo += lineInfoTemp
+				flagErr += flagErrTemp
+			} else {
+				lineInfo += "error0;false;"
+				fmt.Printf("error0, incorrect ft.Mode: %v\n", ft.Mode)
+				flagErr += 1
 			}
 		} else if ft.ID > 0 {
 			// UPDATE
 			result, err := db.ExecContext(ctx, `
 				UPDATE financeTracker 
-				SET dateIn = ?, year = ?, month = ?, day = ?, account = ?, product = ?, priceIntx100 = ?, category = ?,
+				SET dateIn = ?, year = ?, month = ?, day = ?, mode = ?, account = ?, product = ?, priceIntx100 = ?, category = ?,
 					commentInt = ?, commentString = ?, checked = ?, dateChecked = ?, exported = 0
 				WHERE ID = ?
 					AND gofiID = ?;
 				`,
-				ft.Date, ft.DateDetails.Year, ft.DateDetails.Month, ft.DateDetails.Day, ft.Account, ft.Product, ft.PriceIntx100, ft.Category,
+				ft.Date, ft.DateDetails.Year, ft.DateDetails.Month, ft.DateDetails.Day, ft.Mode, ft.Account, ft.Product, ft.PriceIntx100, ft.Category,
 				ft.CommentInt, ft.CommentString, ft.Checked, ft.DateChecked,
 				ft.ID, ft.GofiID,
 			)
@@ -340,13 +367,105 @@ func ImportCSV(ctx context.Context, db *sql.DB,
 					if rows != 1 {
 						lineInfo += "unknown ID;false;"
 					} else {
-						lineInfo += ";true;"
+						if ft.Mode == 0 {
+							iDlistInt = append(iDlistInt, ft.ID)
+							lineInfo += ";true;"
+						} else if ft.Mode >= 1 && ft.Mode <= 4 {
+							lb.FT = ft
+							lb.ModeInt = ft.Mode
+							lb.CreateLenderBorrowerName = row[7]
+							lb.Who = row[7]			
+							lineInfoTemp, flagErrTemp = handleLBinsertCSV(ctx, &lb, csvSeparator, csvDecimalDelimiter, dateFormat, dateSeparator)
+							lineInfo += lineInfoTemp
+							flagErr += flagErrTemp
+							lineInfo += ";true;"
+						} else {
+							iDlistInt = append(iDlistInt, ft.ID)
+							lineInfo += "error5;false;"
+							fmt.Printf("error5, incorrect ft.Mode: %v\n", ft.Mode)
+							flagErr += 1
+						}
 					}
 				}
 			}
 		}
 		stringList += lineInfo + "\n"
 	}
+	b := DeleteSpecificRecordsByMode(ctx, appdata.DB, gofiID, &iDlistInt)
+	if b {
+		fmt.Println("DeleteSpecificRecordsByMode error in csv")
+	}
+	b = UpdateRowsInFinanceTrackerToMode0(ctx, appdata.DB, gofiID, &iDlistInt)
+	if b {
+		fmt.Println("UpdateRowsInFinanceTrackerToMode0 error in csv")
+	}
 	stringList = "erreurs rencontrées: " + strconv.Itoa(flagErr) + "\n" + stringList
 	return stringList, errorBool
+}
+
+func handleLBinsertCSV(ctx context.Context, lb *appdata.LendBorrow,
+	csvSeparator rune, csvDecimalDelimiter string, dateFormat string, dateSeparator string) (string, int) {
+	// ft.Mode >= 1 && ft.Mode <= 4 {
+	var lineInfo string = ""
+	var flagErr int = 0
+	isErr := InsertUpdateInLenderBorrower(ctx, appdata.DB, lb) // 1.
+	if isErr {
+		lineInfo += "handleLBinsertCSV-error1;false;"
+		fmt.Println("handleLBinsertCSV-error1")
+		flagErr += 1
+	} else {
+		var idFT int64
+		var isErr bool
+		var info string
+		if lb.FT.ID == 0 {
+			idFT, isErr, _, info = handleFTinsertCSV(ctx, &lb.FT, csvSeparator, csvDecimalDelimiter, dateFormat, dateSeparator) // 2.
+		} else {
+			idFT = int64(lb.FT.ID)
+			isErr = false
+		}
+		if isErr {
+			lineInfo += "handleLBinsertCSV-error2;false;"
+			fmt.Printf("handleLBinsertCSV-error2: %v\n", info)
+			flagErr += 1
+		} else {
+			lb.FT.ID = int(idFT)
+			isErr = InsertInSpecificRecordsByMode(ctx, appdata.DB, lb) // 3.
+			if isErr {
+				lineInfo += "handleLBinsertCSV-error3;false;"
+				fmt.Println("handleLBinsertCSV-error3")
+				flagErr += 1
+			} else {
+				lineInfo += strconv.FormatInt(idFT, 10) + ";true;"
+			}
+		}
+	}
+	return lineInfo, flagErr
+}
+
+func handleFTinsertCSV(ctx context.Context, ft *appdata.FinanceTracker, 
+	csvSeparator rune, csvDecimalDelimiter string, dateFormat string, dateSeparator string) (int64, bool, int, string) {
+	if (ft.Mode == 1 || ft.Mode == 4) && ft.PriceDirection == "expense" {
+		fmt.Println("handleFTinsert error4 invalid PriceDirection and Mode combinaison")
+		return 0, true, 0, "invalid PriceDirection and Mode combinaison"
+	}
+	if (ft.Mode == 2 || ft.Mode == 3) && ft.PriceDirection == "gain" {
+		fmt.Println("handleFTinsert error5 invalid PriceDirection and Mode combinaison")
+		return 0, true, 0, "invalid PriceDirection and Mode combinaison"
+	}
+	if ft.PriceDirection == "" {
+		fmt.Println("handleFTinsert error6 invalid PriceDirection")
+		return 0, true, 0, "invalid PriceDirection"
+	} else if ft.PriceDirection == "expense" {
+		ft.FormPriceStr2Decimals = "-" + ft.FormPriceStr2Decimals
+	}
+	if ft.DateChecked == "" {
+		ft.DateChecked = "9999-12-31"
+	}
+	ft.PriceIntx100 = ConvertPriceStrToInt(ft.FormPriceStr2Decimals, csvDecimalDelimiter) 
+	idInserted, err := InsertRowInFinanceTracker(ctx, appdata.DB, ft)
+	if err != nil { // Always check errors even if they should not happen.
+		fmt.Printf("handleFTinsert error sqlite.InsertRowInFinanceTracker: %v\n", err.Error())
+		return 0, true, 0, "server error"
+	}
+	return idInserted, false, 0, ""
 }