+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
 ## v1 rework with go-chi
+- 2024-11-06 : 
+    - feat: handle the update of row with a thirdParty in .csv
+    - fix: put back the user id in lenderBorrower table
+- 2024-11-05 : 
+    - feat: handle the creation of thirdParty with the .csv export
+    - feat: delete the thirdParty by default on .csv update row with mode 0 (or delete which is always mode 0)
+- 2024-11-03 : 
+    - feat: add 2 new columns in the csv
+    - feat: handle the creation of thirdParty with the .csv import
+    - tests: improve .csv tests
+    - refacto: more use of the func handleFTinsert
 - 2024-11-01 : .csv export add a `.` at the end of each row in a dedicated column (handle LF and CRLF EOL cases)
 - 2024-10-27 : 
     - fix: remove quirks mode on most pages and adjust padding